PLAT-773: move folder
